js/xpconnect/tests/mochitest/test_bug803730.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=803730
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 803730</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=803730">Mozilla Bug 803730</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 803730 **/
var foo = {
  isNode: function(obj) {
    return !!(obj instanceof Node);
  }
};

var elem = document.createElement("span");
var trueCount = 0,
    falseCount = 0;
for (var x = 0; x < 100000; x++) {
  if (foo.isNode(elem))
    trueCount++;
  else
    falseCount++;
}
is(falseCount, 0, "elem instanceof Node working correctly.");

</script>
</pre>
</body>
</html>
